job_update_service/requirements.txt
requests
apscheduler
python-dotenv
beautifulsoup4